remove calculators
