+++ lib/cloud_controller/deployment_updater/actions/cancel.rb
@@ -1,4 +1,3 @@
-require 'cloud_controller/deployment_updater/calculators/find_interim_web_process'
 module VCAP::CloudController
   module DeploymentUpdater
     module Actions
@@ -19,7 +18,7 @@ module VCAP::CloudController
 
             deploying_web_process.lock!
 
-            prior_web_process = Calculators::FindInterimWebProcess.new(deployment).call || app.oldest_web_process
+            prior_web_process = find_interim_web_process || app.oldest_web_process
             prior_web_process.lock!
 
             prior_web_process.update(instances: deployment.original_web_process_instance_count)
@@ -32,6 +31,31 @@ module VCAP::CloudController
               status_reason: DeploymentModel::CANCELED_STATUS_REASON
             )
           end
+        end
+
+        private
+
+        def find_interim_web_process
+          # Find newest interim web process that (a) belongs to a SUPERSEDED (DEPLOYED) deployment and (b) has at least
+          # one running instance.
+          app.web_processes_dataset.
+            qualify.
+            join(:deployments, deploying_web_process_guid: :guid).
+            where(deployments__state: DeploymentModel::DEPLOYED_STATE).
+            where(deployments__status_reason: DeploymentModel::SUPERSEDED_STATUS_REASON).
+            order(Sequel.desc(:created_at), Sequel.desc(:id)).
+            find { |p| has_running_instance?(p) }
+        end
+
+        def has_running_instance?(process)
+          instances = instance_reporters.all_instances_for_app(process)
+          instances.any? { |_, val| val[:state] == VCAP::CloudController::Diego::LRP_RUNNING }
+        rescue CloudController::Errors::ApiError # the instances_reporter re-raises InstancesUnavailable as ApiError
+          false
+        end
+
+        def instance_reporters
+          CloudController::DependencyLocator.instance.instances_reporters
         end
       end
     end
