+++ lib/cloud_controller/deployment_updater/actions/scale.rb
@@ -1,7 +1,6 @@
 require 'cloud_controller/deployment_updater/actions/scale_down_canceled'
 require 'cloud_controller/deployment_updater/actions/scale_down_old_process'
 require 'cloud_controller/deployment_updater/actions/finalize'
-require 'cloud_controller/deployment_updater/calculators/all_instances_routable'
 
 module VCAP::CloudController
   module DeploymentUpdater
@@ -18,7 +17,7 @@ module VCAP::CloudController
         def call
           deployment.db.transaction do
             return unless deployment.lock!.state == DeploymentModel::DEPLOYING_STATE
-            return unless Calculators::AllInstancesRoutable.new(deployment, logger).call
+            return unless all_instances_routable?
 
             app.lock!
             oldest_web_process_with_instances.lock!
@@ -59,6 +58,18 @@ module VCAP::CloudController
 
         def deploying_web_process
           @deploying_web_process ||= deployment.deploying_web_process
+        end
+
+        def all_instances_routable?
+          instances = instance_reporters.all_instances_for_app(deployment.deploying_web_process)
+          instances.all? { |_, val| val[:state] == VCAP::CloudController::Diego::LRP_RUNNING && val[:routable] }
+        rescue CloudController::Errors::ApiError # the instances_reporter re-raises InstancesUnavailable as ApiError
+          logger.info("skipping-deployment-update-for-#{deployment.guid}")
+          false
+        end
+
+        def instance_reporters
+          CloudController::DependencyLocator.instance.instances_reporters
         end
       end
     end
